requirements.txt
uvicorn[standard]
fastapi
python-dotenv
pydantic-settings
requests